+++ src/open-api-helpers.ts
@@ -561,8 +561,10 @@ export function extractObjectSchema(subject: IOSchema) {
     objectSchema = Array.from(subject.options.values())
       .map((option) => extractObjectSchema(option))
       .reduce((acc, option) => acc.merge(option.partial()), z.object({}));
+  } else if (subject instanceof z.ZodEffects) {
+    objectSchema = extractObjectSchema(subject._def.schema); // object refinement
   } else {
-    // intersection schema
+    // intersection
     objectSchema = extractObjectSchema(subject._def.left).merge(
       extractObjectSchema(subject._def.right)
     );