+++ CHANGELOG.md
@@ -2,6 +2,34 @@
 
 ## Version 8
 
+### v8.3.0
+
+- Feature #600: Top level refinements.
+  - Starting this version you can use the `.refine()` method on the `z.object()` of the input schema.
+  - This feature might be useful, for example, when you have multiple optional properties on the top level, but at
+    least one of them has to be specified.
+  - Currently, only the refinements of `z.object()` are supported.
+    - You can not combine it with `z.union()`, `z.intersetion()`, `z.discriminatedUnion()`, `.or()`, `.and()` yet.
+
+```typescript
+// example
+import { z } from "express-zod-api";
+
+const endpoint = endpointsFactory.build({
+  input: z
+    .object({
+      email: z.string().email().optional(),
+      id: z.string().optional(),
+      otherThing: z.string().optional(),
+    })
+    .refine(
+      (inputs) => Object.keys(inputs).length >= 1,
+      "Please provide at least one property"
+    ),
+  // ...
+});
+```
+
 ### v8.2.1
 
 - OpenAPI generator throws when attempting to use `z.file()` within input schema.
@@ -502,7 +530,6 @@ export const myMiddleware = createMiddleware({
     .passthrough(), // <— remove this if you have it in your code
   middleware: async () => ({...}),
 });
-
 ```
 
 ## Version 5